Dev/remove ctx (#9)
* remove ctx

* gofmt

* add workflows

* add code

* add code

* add code

* add code

* fix code yml

* update workflows go version

* add go test debug tags

* optimize ut

* optimize ut

* fix ut

* fix ut

* fix ut

* add ci test

* add code

* add code

+++ helper/bpf/bpf_test.go
@@ -38,7 +38,7 @@ func Test_FilterDstPortAndSynFlag(t *testing.T) {
 
 		n, err := vm.Run(e.IP)
 		require.NoError(t, err)
-		require.Equal(t, e.Ret, n)
+		require.Equal(t, e.Ret, uint16(n))
 
 	}
